+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Features:
   a separate bundle in this repo.
 * New console command `mapbender:normalize-translations` to quickly find and complement missing translations ([PR#1538](https://github.com/mapbender/mapbender/pull/1538))
 * New console command `mapbender:wms:assign` to add a wms source instance to an application ([PR#1552](https://github.com/mapbender/mapbender/pull/1552))
-* Modified command `mapbender:wms:show`. Command id is not optional, if omitted all sources are shown. Also, id can be replaced by origin url. ([PR#1552](https://github.com/mapbender/mapbender/pull/1552))
+* Modified command `mapbender:wms:show`. Command id is not optional, if omitted all sources are shown; id can be replaced by origin url; add option `--json` for json output. ([PR#1552](https://github.com/mapbender/mapbender/pull/1552))
 * [SearchRouter] New option exportcsv to download the result list as CSV ([PR#1509](https://github.com/mapbender/mapbender/pull/1509))
 * [ApplicationAssetService] Allow overriding sass/css and js assets by calling ApplicationAssetService::registerAssetOverride or by using the new parameter `mapbender.asset_overrides` ([PR#1512](https://github.com/mapbender/mapbender/pull/1512))
 * [Button] Allow customization of the icons available for selection in the button edit form. See PR description for details. ([PR#1518](https://github.com/mapbender/mapbender/pull/1518))